@@ -8,13 +8,13 @@ class ReactionsHelperTest < ActiveSupport::TestCase
   test "status to integer index works" do
     (1..4).each do |n|
       res = status_to_integer_index(n)
-      assert res.class == Fixnum
+      assert res.class == Integer
       assert n == (res + 1)
     end
 
     (1..4).each do |n|
       res = status_to_integer_index(n.to_s)
-      assert res.class == Fixnum
+      assert res.class == Integer
       assert n == (res + 1)
     end
     
